migrate get method to own file, rename test, add coverage
